+++ LookupAnything/release-notes.md
@@ -5,6 +5,7 @@
 * Updated for Stardew Valley 1.4, including new content.
 * Fixed player luck in multiplayer showing your own luck instead of theirs.
 * Fixed 'scissor rectangle is invalid' error in rare cases.
+* Fixed game freeze if you open a lookup on the load screen and then close it by pressing `F1` again.
 * Improved translations. Thanks to shiro2579 (updated Portuguese)!
 
 ## 1.22.2